refactor: 패키지 이동, 불필요한 코드 제거 #15
