+++ src/main/java/com/munecting/api/domain/user/dao/UserRepository.java
@@ -1,6 +1,5 @@
-package com.munecting.api.domain.user.repository;
+package com.munecting.api.domain.user.dao;
 
-import com.munecting.api.domain.user.dto.SocialType;
 import com.munecting.api.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,9 +8,6 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
-    Boolean existsByEmail(String email);
-    Boolean existsByNickname(String nickname);
-
     Optional<User> findByEmail(String email);
 
 }
