--- src/main/java/com/munecting/api/global/Redis/repository/RefreshRepository.java
@@ -1,10 +1,0 @@
-package com.munecting.api.global.Redis.repository;
-
-import com.munecting.api.global.Redis.entity.RefreshEntity;
-import org.springframework.data.repository.CrudRepository;
-
-
-public interface RefreshRepository extends CrudRepository<RefreshEntity, String> {
-
-
-}
